ui: shared: stats on dashboard
