+++ src/App.css
@@ -157,8 +157,8 @@ body {
 
 /* ######### components ######### */
 
-/* home */
-.home-stats-grid {
+/* shared */
+.stats-grid {
     display: grid;
     grid-gap: 1.5rem;
     grid-template-columns: repeat(auto-fit, minmax(8rem, 1fr));
