+++ src/comps/home/Home.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { fm } from '../../data/utils'
 import mj from '../../res/mj.png'
+import StatsWidget from '../shared/StatsWidget'
 
 export default props =>
   <React.Fragment>
@@ -19,34 +19,7 @@ export default props =>
     </div>
 
     {/* Majorna stats from firestore/mj/meta document */}
-    {!props.mjMetaDoc ? <div className="mj-box center-all spinner"/> :
-      <div className="mj-box home-stats-grid">
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-success">${props.mjMetaDoc.val}</strong>
-          Current Price
-        </div>
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-success">${fm(props.mjMetaDoc.marketCap * props.mjMetaDoc.val)}</strong>
-          Market Cap.
-        </div>
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-success">${fm(props.mjMetaDoc.monthly.txVolume * props.mjMetaDoc.val)}</strong>
-          Tx Volume (monthly)
-        </div>
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-grey">{fm(props.mjMetaDoc.marketCap / 1000000)}M mj</strong>
-          Current Supply
-        </div>
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-grey">{fm(props.mjMetaDoc.maxSupply / 1000000)}M mj</strong>
-          Max Supply
-        </div>
-        <div className="flex-column center-all">
-          <strong className="is-size-5 has-text-grey">{props.mjMetaDoc.userCount}</strong>
-          Total Accounts
-        </div>
-      </div>
-    }
+    {!props.mjMetaDoc ? <div className="mj-box center-all spinner"/> : <StatsWidget mjMetaDoc={props.mjMetaDoc} />}
 
     {/* Media items with key Majorna facts */}
     <div className="mj-box align-start">
